+++ digest/src/core_api/wrapper.rs
@@ -3,7 +3,8 @@ use super::{
     UpdateCore, XofReaderCoreWrapper,
 };
 use crate::{
-    ExtendableOutput, ExtendableOutputReset, FixedOutput, FixedOutputReset, HashMarker, Update,
+    CustomizedInit, ExtendableOutput, ExtendableOutputReset, FixedOutput, FixedOutputReset,
+    HashMarker, Update,
 };
 use block_buffer::BlockBuffer;
 use core::{
@@ -160,8 +161,28 @@ impl<T: BufferKindUser> Drop for CoreWrapper<T> {
 #[cfg(feature = "zeroize")]
 impl<T: BufferKindUser + zeroize::ZeroizeOnDrop> zeroize::ZeroizeOnDrop for CoreWrapper<T> {}
 
+impl<T> CustomizedInit for CoreWrapper<T>
+where
+    T: BufferKindUser + CustomizedInit,
+{
+    type CustomizedExtArg = T::CustomizedExtArg;
+
+    #[inline]
+    fn new_customized(customization: &[u8]) -> Self {
+        Self::from_core(T::new_customized(customization))
+    }
+
+    #[inline]
+    fn new_ext_customized(customization_ext: &Self::CustomizedExtArg) -> Self {
+        Self::from_core(T::new_ext_customized(customization_ext))
+    }
+}
+
 #[cfg(feature = "oid")]
-impl<T: BufferKindUser + AssociatedOid> AssociatedOid for CoreWrapper<T> {
+impl<T> AssociatedOid for CoreWrapper<T>
+where
+    T: BufferKindUser + AssociatedOid,
+{
     const OID: ObjectIdentifier = T::OID;
 }