+++ digest/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## UNRELEASED
+### Added
+- `CustomizedInit` trait ([#1334]).
+
 ### Changed
 - `crypto-common` dependency bumped to v0.2 ([#1173])
 - Edition changed to 2021 and MSRV bumped to 1.57 ([#1173])
@@ -14,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Mac::new`, `Mac::new_from_slice`, and `Mac::generate_key` methods ([#1173])
 
 [#1173]: https://github.com/RustCrypto/traits/pull/1173
+[#1334]: https://github.com/RustCrypto/traits/pull/1334
 
 ## 0.10.7 (2023-05-19)
 ### Changed